added few comments
